Create SSR and client validators
