+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "clean": "rm -rf lib/*",
     "lint": "eslint src",
-    "test": "cross-env NODE_ENV=test karma start --single-run",
+    "test": "jest && cross-env NODE_ENV=test karma start --single-run",
     "test:watch": "cross-env NODE_ENV=test karma start",
     "prebuild": "npm run clean && npm run lint && npm test",
     "build": "babel src --out-dir lib --ignore tests",
@@ -64,6 +64,7 @@
     "mocha": "^3.3.0",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
+    "react-html-attributes": "^1.3.0",
     "requirejs": "^2.3.3",
     "svg-elements": "^1.0.2",
     "webpack": "^2.4.1"
@@ -76,5 +77,11 @@
   "dependencies": {
     "karma": "^1.6.0",
     "ramda": "^0.23.0"
+  },
+  "jest": {
+    "rootDir": "src",
+    "testPathIgnorePatterns": [
+      "client"
+    ]
   }
 }
