+++ src/tests/ssr/validAttr.test.js
@@ -1,0 +1,6 @@
+import validAttrTests from '../validAttrTests'
+import { validAttr } from '../../ssr'
+
+describe('validAttr', () => {
+  validAttrTests(validAttr)
+})
